Added Rumble Command
This command causes the driver controller to rumble for a given amount of time. As of right now, the only time this is used is if the operator tries to set the arm to the high position when we have a tilted cone, since we are not able to score tilted cones on the high nodes.